requirements.txt
django
djangorestframework
djangorestframework-jwt
django-cors-headers
six
cryptography
django-sslserver
